Implement process scheduling
Provides an initial attempt at implementing the core logic for
scheduling processes. We have yet to add proper printing functionality
so that the users can follow what the scheduler does at each step, and
also verify with a few examples that the scheduling logic is free of
bugs (and complies with what was specified).